requirements.txt
dash-cytoscape==1.0.1
mysql-connector-python>=8.0.0